% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass-attic.R
\name{add_adiff}
\alias{add_adiff}
\title{Add adiff attribute to overpass query}
\usage{
add_adiff(opq, olddatetime, newdatetime)
}
\arguments{
\item{opq}{An`overpass_query` object}

\item{olddatetime}{datetimestring in the form %Y-%m-%dT%H:%M:%SZ e.g. 2019-01-01T12:00:00Z}

\item{newdatetime}{optional: datetimestring in the form %Y-%m-%dT%H:%M:%SZ e.g. 2019-01-01T12:00:00Z, if empty current time is used}
}
\value{
osmdiff object
}
\description{
Add adiff attribute to overpass query
}
\examples{
\dontrun{
oq <- osmdata::opq('bhaktapur')
add_date(oq,datetime)
}
}
\references{
\url{https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#date}
}
