% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass-attic.R
\name{add_date}
\alias{add_date}
\title{Add date attribute to overpass query}
\usage{
add_date(opq, datetime)
}
\arguments{
\item{opq}{An`overpass_query` object}

\item{datetime}{datetimestring in the form 2019-01-01T12:00:00Z}
}
\value{
osmtt object
}
\description{
Add date attribute to overpass query
}
\examples{
oq <- osmdata::opq('bhaktapur')
add_date(oq,"2019-01-01T12:00:00Z")
}
