% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmtt.R
\name{osmadiff}
\alias{osmadiff}
\title{Get how the osm data looked at certain time}
\usage{
osmadiff(bbox, datetime)
}
\arguments{
\item{bbox}{place name or bbox object}

\item{datetime}{The timestamp where you want the data}
}
\description{
Get how the osm data looked at certain time
}
