% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdiff.R
\name{osmdiff}
\alias{osmdiff}
\title{osmdiff class def}
\usage{
osmdiff(bbox = NULL, file = NULL,
  base_url = "http://overpass-api.de/api", silent = FALSE)
}
\arguments{
\item{bbox}{bounding box}

\item{file}{previous export}

\item{base_url}{change if required, default is "http://overpass-api.de/api"}

\item{silent}{produce warning messages}

\item{...}{other options ignored}
}
\description{
osmdiff class def
}
\note{
Class constructor should never be used directly, and is only exported
to provide access to the print method
}
