% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmtt-methods.R
\name{osmtt_sf}
\alias{osmtt_sf}
\title{queries overpass object and convert to osmdata object}
\usage{
osmtt_sf(q)
}
\arguments{
\item{q}{osmtt object}
}
\value{
osmdata object
}
\description{
queries overpass object and convert to osmdata object
}
\examples{
\dontrun{
new_url <- "http://overpass-api.de/api/interpreter"
osmdata::set_overpass_url(new_url)
op <- osmtt("sudal nepal","2019-01-01T12:00:00Z")
op_sf <- osmtt_sf(op,quiet = FALSE)
}
}
