% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmttdiff-methods.R
\name{osmttdiff_sf}
\alias{osmttdiff_sf}
\title{Converts overpass diff xml to osmttdiff object}
\usage{
osmttdiff_sf(xml)
}
\arguments{
\item{xml}{xml document from overpass diff query}
}
\description{
Converts overpass diff xml to osmttdiff object
}
\examples{
\dontrun{
(od<-osmadiff("sanga","2019-01-01T12:00:00Z"))
odd<-osmdata:::overpass_query(osmdata::opq_string(od))
osmttdiff_sf(odd) -> odd_sf
}
}
