% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass-attic.R
\name{add_adiff}
\alias{add_adiff}
\title{Add adiff attribute to overpass query}
\usage{
add_adiff(opq, olddatetime, newdatetime)
}
\arguments{
\item{opq}{An overpass_query object}

\item{olddatetime}{datetimestring in the form 2019-01-01T12:00:00Z}

\item{newdatetime}{optional datetimestring in the form 2019-01-01T12:00:00Z, if empty current time is used}
}
\value{
osmdiff object
}
\description{
Add adiff attribute to overpass query
}
\examples{
oq <- osmdata::opq('bhaktapur')
add_adiff(oq,"2019-01-01T12:00:00Z","2019-08-01T12:00:00Z")
}
