% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmtt.R
\name{add_date}
\alias{add_date}
\title{Add date attribute to overpass query}
\usage{
add_date(opq, datetime)
}
\arguments{
\item{opq}{An`overpass_query` object}

\item{datetime}{datetimestring in the form %Y-%m-%dT%H:%M:%SZ e.g. 2019-01-01T12:00:00Z}
}
\value{
\link{opq} object
}
\description{
Add date attribute to overpass query
}
\examples{
\dontrun{
oq <- osmdata::opq('bhaktapur')
add_date(oq,datetime)
}
}
\references{
<https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#date>
}
