% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmtt.R
\name{osmtt}
\alias{osmtt}
\title{Get how the osm data looked at past time}
\usage{
osmtt(bbox, datetime)
}
\arguments{
\item{bbox}{place name or bbox object}

\item{datetime}{The timestamp where you want the data}
}
\description{
Get how the osm data looked at past time
}
\examples{
plot(osmtt("muenster",""))
}
